backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
faiss-cpu==1.7.4
numpy==1.26.4
streamlit
requests
